Make RT owner dropdown limit a config option
The owner dropdown menu on a ticket's Modify People page becomes an
autocomplete field once there are greater than 50 values. As some
users may want to change this limit, a new config option was added
so it can easily be updated, rather than using a callback. Existing
callbacks on this limit will continue to work as expected.